[OS] 增加errno.h

+++ include/os/type.h
@@ -2,13 +2,13 @@
 #define _TYPE_H_
 typedef unsigned char BYTE;
 typedef unsigned short WORD;
-typedef unsigned long DWORD;
+typedef unsigned int DWORD;
 typedef unsigned char	u8;
 typedef unsigned short	u16;
-typedef unsigned long	u32;
+typedef unsigned int	u32;
 typedef signed char	s8;
 typedef signed short	s16;
-typedef signed long	s32;
+typedef signed int	s32;
 
 #ifndef _SIZE_T
 #define _SIZE_T
